Fixed Comments in PR
