+++ Recrutify/Recrutify.Services/Services/PrimarySkillService.cs
@@ -24,14 +24,5 @@ namespace Recrutify.Services.Services
             var skills = await _primarySkillRepository.GetAllAsync();
             return _mapper.Map<List<PrimarySkillDTO>>(skills);
         }
-        
-        public IQueryable<PrimarySkillDTO> Get()
-        {
-            var projects = _primarySkillRepository.Get();
-            //var arr1 = projects.ToArray();
-            var dto = _mapper.ProjectTo<PrimarySkillDTO>(projects);
-            //var arr2 = dto.ToArray();
-            return dto;
-        }
     }
 }
